Added sum 2D

+++ fpb/sum/numpy.py
@@ -3,5 +3,5 @@ from fpb.base import numpy
 
 class Runner(numpy.BaseNumpy1dRunner):
     def run(self, data):
-        output = self.np.sum(data)
+        output = self.np.sum(data, dtype="float128")
         return output
